Reuse legend code in both renderers
